Squashed commit of the following:
commit 041f833c925fc875d4b455bcbddd645800e3f2cf
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 22 10:57:36 2024 +0900

    docs: 1.0 (7) 테플버전

commit 470e9c5930be2f6a60d999d21b124bdb0e8e3f8e
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 22 10:44:47 2024 +0900

    fix: 대댓글 달면 내가 대댓글 단 댓글은 펼쳐 보여지도록 수정

commit b34cedd487622a79a11939a253e006b36a88312b
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 22 10:01:19 2024 +0900

    fix: 업로드시 이미지의 정확한 사이즈를 캡쳐하지 못해서 비율 망가지는 문제 수정

commit 402bc0d1842e1937efb5b2f2519361c5bd85533a
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 22 10:00:49 2024 +0900

    feat: 인앱리뷰 기능 추가

commit fcd209f051a54afa9b3a0f12ead60a67a999998a
Author: dudu <whqtkf12@naver.com>
Date:   Wed Feb 21 14:43:43 2024 +0900

    feat: 회원가입 실패시, 이유 알려주도록 수정

commit 3302fb52591d118748a9d6da33541f1cbdf1baef
Author: dudu <whqtkf12@naver.com>
Date:   Wed Feb 21 14:26:44 2024 +0900

    refactor: 약관 동의 화면에서, Row 터치시 Toggle되는걸로 변경

commit 162bb67e04a04c16b47a80f236ff1f0d613a5ec8
Author: dudu <whqtkf12@naver.com>
Date:   Fri Feb 16 08:31:00 2024 +0900

    fix: 알림 여부에따라 아이콘 다르게 보여주도록 구현 (있을때만 빨간점..!)

commit e00047d8a87a5437bcb7800b17ea5458e82872e3
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 16:48:56 2024 +0900

    fix: 살말 업로드시, 필터링된 이미지 가끔 90도 회전해 있는 문제 수정

commit c5758edff804c00f129b062cca4228aecca4e839
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 16:48:30 2024 +0900

    fix: 살말 업로드시, 확인/취소 버튼 간헐적으로 눌리지 않는 문제 수정

commit 617b0f89954c65c504a3dbf4fb49cb975fd08064
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 15:23:26 2024 +0900

    fix: 이미지 업로드시 간헐적으로 취소, 확인버튼 눌리지 않는 문제 수정

commit cd6a0434c978b06ddd644c3311cb769726eaaadd
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 14:32:35 2024 +0900

    fix: 사진편집화면에서 비율 맞지 않는 문제 수정

commit 518b2fd971d1ece010b71dd2fb349f2cddf78ef5
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 13:24:10 2024 +0900

    fix: 살말 업로드시 카메라 사용후 crash / 무한 모달 현상 수정

commit 60786b1bd190584d78fcf0819749ff5d4bfbeb09
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 13:23:49 2024 +0900

    doc: 빌드버전 4->5로 변경

commit 9e02029408632c46262a05a248cd39ede7723ab3
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 11:01:19 2024 +0900

    fix: 살말 하나도 없을때 투표버튼 누르면 터지는 현상 수정

commit 7ea061ccee842f9d002b2a0bd02a83f4de19243e
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 10:54:23 2024 +0900

    fix: 업로드한 살말 삭제시, 해당 타켓이 제대로 삭제되도록 수정

commit 1cedce559de728bc1a435f4b94732267a965b6c8
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 10:53:58 2024 +0900

    fix: 살말 업로드시 홈탭 업데이트 되도록 수정

commit 4aa6247292601db719a33016ac0e2ac1cde3c397
Author: dudu <whqtkf12@naver.com>
Date:   Thu Feb 15 10:53:01 2024 +0900

    fix: 약관전체동의 / 개별동의 toggle에 따라 버튼 enable / disable 제대로 동작하도록 수정

commit dabc87c049d9f388f9bb3b3b2e1b974a57787466
Author: dudu <whqtkf12@naver.com>
Date:   Wed Feb 14 13:59:36 2024 +0900

    feat: Swipe to Back 기능 추가

+++ Projects/ProfileFeature/Sources/Views/UploadList/UploadListCore.swift
@@ -7,6 +7,7 @@ public struct UploadListCore: Reducer {
 
   public struct State: Equatable {
     @BindingState var isDeletionPresented: Bool = false
+    var targetVote: Vote?
 
     var votes: IdentifiedArrayOf<Vote> = []
 
@@ -17,8 +18,9 @@ public struct UploadListCore: Reducer {
     case binding(BindingAction<State>)
     case dismissButtonTapped
     case confirmButtonTapped
-    case removeVoteTapped(Vote)
-
+    case deleteButtonTapped
+    case remoteButtonTapped(Vote)
+    
     case _onAppear
     case _removeVoteResponse(TaskResult<Int>)
 
@@ -39,7 +41,7 @@ public struct UploadListCore: Reducer {
       switch action {
       case .binding:
         return .none
-
+        
       case .dismissButtonTapped:
         return .run { send in
           await dismiss()
@@ -47,9 +49,16 @@ public struct UploadListCore: Reducer {
 
       case .confirmButtonTapped:
         return .none
+        
+      case let .remoteButtonTapped(vote):
+        state.isDeletionPresented = true
+        state.targetVote = vote
+        return .none
 
-      case .removeVoteTapped(let vote):
-        return .run { [id = vote.id] send in
+      case .deleteButtonTapped:
+        guard let id = state.targetVote?.id else { return .none }
+        
+        return .run { send in
           await send(._removeVoteResponse(
             TaskResult {
               try await voteRepository.delete(voteID: id)
@@ -61,6 +70,7 @@ public struct UploadListCore: Reducer {
       case ._onAppear:
         return .run { send in
           let votes = try await memberRepository.votes(memberID: userDefault.memberID ?? -1, cursorId: nil, size: 100)
+          
           await send(._setVotes(votes))
         }
